requirements.txt
Flask==2.0.1
Flask-Uploads==0.2.1
Pillow==8.3.2
numpy==1.21.1
tensorflow==2.5.0